requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.2.0
httpx==0.27.2
aiofiles==24.1.0
Pillow==10.4.0

# AWS S3
boto3==1.35.0
botocore==1.35.0

# Redis + Celery
redis==5.0.8
celery[redis]==5.4.0
flower==2.0.1

# WebSocket
websockets==13.0

# Rate limiting & security
slowapi==0.1.9
itsdangerous==2.2.0
cryptography==43.0.1

# Idempotency / caching
cachetools==5.5.0

# 2FA
pyotp==2.9.0

# Structured logging
structlog==24.4.0
